Started on font viewer

+++ Editor/src/Panels/ViewportPanel.cpp
@@ -644,7 +644,7 @@ void ViewportPanel::OnImGuiRender()
 
 		ImVec2 window_pos = ImGui::GetCursorScreenPos();
 
-		uint64_t tex = (uint64_t)m_Framebuffer->GetColourAttachment();
+		uintptr_t tex = (uintptr_t)m_Framebuffer->GetColourAttachment();
 
 		ImGui::Image((void*)tex, m_ViewportSize, ImVec2(0, 1), ImVec2(1, 0));
 		ImGui::PopStyleVar(2);
@@ -764,7 +764,7 @@ void ViewportPanel::OnImGuiRender()
 							cameraPreviewPosition.y + m_CameraPreview->GetSpecification().height + 24.0f),
 						color, ImGui::GetStyle().WindowRounding);
 
-					uint64_t cameraTex = (uint64_t)m_CameraPreview->GetColourAttachment();
+					uintptr_t cameraTex = (uintptr_t)m_CameraPreview->GetColourAttachment();
 					float cameraCursorPosition = topLeft.x - ImGui::GetStyle().ItemSpacing.x + m_ViewportSize.x - m_CameraPreview->GetSpecification().width;
 					ImGui::SetCursorPos(ImVec2(cameraCursorPosition, topLeft.y - ImGui::GetStyle().ItemSpacing.y + m_ViewportSize.y - m_CameraPreview->GetSpecification().height - 21.0f));
 					ImGui::Text(" %s", selectedEntity.GetName().c_str());
